requirements.txt
matplotlib~=3.8.2
numpy~=1.26.3
torch~=2.1.2+cu121
torchmetrics~=1.3.0.post0
torchvision~=0.16.2+cu121
Pillow~=9.5.0
scikit-learn~=1.4.0
tabulate~=0.9.0
trdg~=1.8.0